requirements.txt
aiogram>=3.3,<4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
redis[hiredis]>=5.0.0
pydantic-settings>=2.1.0
alembic>=1.13.0
python-dotenv>=1.0.0
